refactor rooms and houses api file names to be plural, and add api_users file
