+++ src/dashboard/src/Pages/ManageUsers/NewUserModal.js
@@ -8,8 +8,8 @@ import Backdrop from "@material-ui/core/Backdrop";
 import TextField from "@material-ui/core/TextField";
 import Button from "@material-ui/core/Button";
 import {toast} from "react-toastify";
-import {createNewAgent} from "../../Api/api_agents";
-import {getHouseList} from "../../Api/api_house";
+import {createNewUser} from "../../Api/api_users";
+import {getHouseList} from "../../Api/api_houses";
 import {useHistory} from 'react-router-dom'
 
 const NewUserModal = ({open, onClose, refreshUsers}) => {
